+++ src/test/java/io/github/astrapi69/date/ConvertDateExtensionsTest.java
@@ -68,6 +68,7 @@ public class ConvertDateExtensionsTest
 	@AfterMethod
 	protected void tearDown()
 	{
+		this.format1 = null;
 	}
 
 	/**